Abstract the `translate` function from `trove/parse-pyret.js`.
Instead of having it directly call constructors from `ast.arr`, have
it instead look up the function to call in a dictionary. Furthermore,
move it into `base/translate-parse-tree.js` so that it can easily be
called from outside Pyret.

This way it can be used both by Pyret to construct an `ast.arr` AST,
and by the "codemirror-blocks" block editor to construct a blocky AST.
